Make the routes config dynamically
In the `config/laravel-impersonate.php` the `take_redirect_to` and `leave_redirect_to` are hard coded urls. This is fine as long as you don't construct your routes dynamically.

Since `route()` in a config file leads to a "Route not defined" error, because Laravel loads the configuration before the routes are registered, the only way to tackle this is to expect a route name.